hw-4: Refactoring and 2 timers
