+++ app/src/main/java/com/gb/stopwatch/model/Stopwatch.kt
@@ -1,0 +1,10 @@
+package com.gb.stopwatch.model
+
+import kotlinx.coroutines.flow.StateFlow
+
+interface Stopwatch {
+    val ticker: StateFlow<String>
+    fun start()
+    fun pause()
+    fun stop()
+}
