+++ app/src/main/java/com/gb/stopwatch/model/stopwatchImpl/StopwatchListOrchestrator.kt
@@ -1,0 +1,51 @@
+package com.gb.stopwatch.model.stopwatchImpl
+
+import com.gb.stopwatch.model.Stopwatch
+import com.gb.stopwatch.model.stopwatchImpl.TimestampMillisecondsFormatter.Companion.DEFAULT_TIME
+import kotlinx.coroutines.*
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.StateFlow
+
+class StopwatchListOrchestrator(
+    private val stopwatchStateHolder: StopwatchStateHolder,
+    private val scope: CoroutineScope,
+) : Stopwatch {
+
+    private var job: Job? = null
+    private val mutableTicker = MutableStateFlow(DEFAULT_TIME)
+    override val ticker: StateFlow<String> = mutableTicker
+
+    override fun start() {
+        if (job == null) startJob()
+        stopwatchStateHolder.start()
+    }
+
+    private fun startJob() {
+        job = scope.launch {
+            while (isActive) {
+                mutableTicker.value = stopwatchStateHolder.getStringTimeRepresentation()
+                delay(20)
+            }
+        }
+    }
+
+    override fun pause() {
+        stopwatchStateHolder.pause()
+        stopJob()
+    }
+
+    override fun stop() {
+        stopwatchStateHolder.stop()
+        stopJob()
+        clearValue()
+    }
+
+    private fun stopJob() {
+        scope.coroutineContext.cancelChildren()
+        job = null
+    }
+
+    private fun clearValue() {
+        mutableTicker.value = DEFAULT_TIME
+    }
+}
