+++ app/src/main/java/com/gb/stopwatch/model/stopwatchImpl/StopwatchState.kt
@@ -1,0 +1,6 @@
+package com.gb.stopwatch.model.stopwatchImpl
+
+sealed class StopwatchState {
+    data class Paused(val elapsedTime: Long) : StopwatchState()
+    data class Running(val startTime: Long, val elapsedTime: Long) : StopwatchState()
+}
