+++ app/src/main/java/com/gb/stopwatch/model/stopwatchImpl/StopwatchStateHolder.kt
@@ -1,0 +1,31 @@
+package com.gb.stopwatch.model.stopwatchImpl
+
+class StopwatchStateHolder(
+    private val stopwatchStateCalculator: StopwatchStateCalculator,
+    private val elapsedTimeCalculator: ElapsedTimeCalculator,
+    private val timestampMillisecondsFormatter: TimestampMillisecondsFormatter
+) {
+
+    var currentState: StopwatchState = StopwatchState.Paused(0)
+        private set
+
+    fun start() {
+        currentState = stopwatchStateCalculator.calculateRunningState(currentState)
+    }
+
+    fun pause() {
+        currentState = stopwatchStateCalculator.calculatePausedState(currentState)
+    }
+
+    fun stop() {
+        currentState = StopwatchState.Paused(0)
+    }
+
+    fun getStringTimeRepresentation(): String {
+        val elapsedTime = when (val currentState = currentState) {
+            is StopwatchState.Paused -> currentState.elapsedTime
+            is StopwatchState.Running -> elapsedTimeCalculator.calculate(currentState)
+        }
+        return timestampMillisecondsFormatter.format(elapsedTime)
+    }
+}
