+++ app/src/main/java/com/gb/stopwatch/model/stopwatchImpl/TimestampProvider.kt
@@ -1,0 +1,5 @@
+package com.gb.stopwatch.model.stopwatchImpl
+
+interface TimestampProvider {
+    fun getMilliseconds(): Long
+}
